# public/openapi.yaml
openapi: 3.0.3
info:
  title: Speclint API
  description: |
    Lint your GitHub issues before AI coding agents touch them. Get completeness_score and agent_ready signal for every spec.

    ## Authentication
    Free tier (5 items/request) requires no authentication.
    Pro and Team tiers require a license key passed via the `x-license-key` header.

    ## Rate Limits
    - **Free**: 5 items per request, rate-limited by IP
    - **Pro** ($29/mo): 25 items per request
    - **Team** ($79/mo): 50 items per request
  version: "1.0.0"
  contact:
    url: https://speclint.ai
  license:
    name: Proprietary
servers:
  - url: https://speclint.ai
    description: Production
paths:
...
  /api/refine:
    post:
      summary: Lint and Score Specs (deprecated alias)
      description: |
        **Deprecated.** Use `/api/lint` instead. This endpoint is a maintained alias
        and will continue to work for backward compatibility, but new integrations should
        use `/api/lint`.
      deprecated: true
      operationId: refineItemsAlias
      security:
        - ApiKeyAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefineRequest"
      responses:
        "200":
          description: Successfully linted items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefineResponse"
...
  /api/verify:
    get:
      summary: Verify a Lint Receipt
      operationId: verifyLintReceipt
      description: |
        Retrieve a stored lint receipt by its `lint_id`. Receipts are public and require no authentication.
        Every lint response includes a `lint_id` (e.g. `spl_a1b2c3d4`). Use this endpoint to verify the result
        was genuinely produced by Speclint and was not fabricated by an intermediary agent.

        Receipts are stored for 30 days.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: The lint_id from a previous lint response (e.g. spl_a1b2c3d4)
          example: "spl_a1b2c3d4"
      responses:
        "200":
          description: Lint receipt found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LintReceipt"
        "400":
          description: Missing or invalid id parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Lint receipt not found (expired or invalid id)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-license-key
      description: |
        License key for Pro or Team tier access.
        Obtain at https://speclint.ai/pricing
        Free tier (5 items/request) works without a key.
  schemas:
    IssueInput:
      type: object
      required:
        - title
        - body
      properties:
        title:
          type: string
          description: Issue title
          example: "Add CSV export to reports"
        body:
          type: string
          description: Issue body / description
          example: "Users need to download report data as CSV for offline analysis in Excel."
        labels:
          type: array
          items:
            type: string
          description: Optional issue labels
          example: ["feature", "reports"]
    RefineRequest:
      type: object
      description: |
        Send either `items` (plain strings) or `issues` (structured GitHub issue objects), not both.
      properties:
        items:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 50
          description: |
            Array of raw backlog item strings to refine.
            Max items depends on tier: 5 (free), 25 (Pro), 50 (Team).
            Use this OR `issues`, not both.
          example:
            - "fix the login bug"
            - "users need to be able to export their data"
        issues:
          type: array
          items:
            $ref: "#/components/schemas/IssueInput"
          minItems: 1
          maxItems: 50
          description: |
            Alternative to `items`. Pass structured GitHub issues with title, body, and optional labels.
            Each issue is converted to a formatted string internally before refinement.
            Use this OR `items`, not both.
          example:
            - title: "Add CSV export to reports"
              body: "Users need to download report data as CSV for offline analysis in Excel."
              labels: ["feature", "reports"]
        context:
          type: string
          description: |
            Optional project context to improve the quality and relevance of refinements.
            Example: "B2B SaaS CRM for enterprise sales teams" or "Mobile fitness app for casual runners".
          example: "B2B project management SaaS for engineering teams"
        useUserStories:
          type: boolean
          default: false
          description: |
            When true, formats titles as user stories:
            "As a [user], I want [goal], so that [benefit]"
        useGherkin:
          type: boolean
          default: false
          description: |
            When true, formats acceptance criteria using Gherkin syntax:
            Given [context], When [action], Then [outcome]
        preserve_structure:
          type: boolean
          default: false
          description: |
            When true, each input string is scored as a single item without decomposition.
            Recommended for CI integrations sending one issue body per request. Default: false.
        persona:
          $ref: "#/components/schemas/PersonaInput"
    PersonaInput:
      type: object
      description: |
        Optional. Target user persona for persona-aware spec scoring (Solo/Team tiers).
        When provided, each refined item receives a `persona_alignment` score (0-100)
        and `persona_gaps` listing which concerns are not addressed in the spec.
        Free tier users receive an upgrade_message instead of persona scores.
      required:
        - role
        - cares_about
        - doesnt_care_about
      properties:
        role:
          type: string
          description: The persona's role or job title
          example: "frontend developer at 3-person AI dev firm"
        cares_about:
          type: array
          items:
            type: string
          description: Things this persona prioritizes
          example: ["CI integration", "agent compatibility", "fast iteration"]
        doesnt_care_about:
          type: array
          items:
            type: string
          description: Things this persona does NOT prioritize
          example: ["enterprise compliance", "manual QA"]
    RefineResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RefinedItem"
        lint_id:
          type: string
          description: Verifiable receipt ID for this lint request. Use GET /api/verify?id={lint_id} to retrieve the stored result.
          example: "spl_a1b2c3d4"
        scores:
          type: array
          description: Completeness scores and agent_ready signals for each item
          items:
            $ref: "#/components/schemas/ItemScore"
        summary:
          $ref: "#/components/schemas/ScoreSummary"
        _meta:
          $ref: "#/components/schemas/ResponseMeta"
        upgrade_message:
          type: string
          nullable: true
          description: |
            Present only when a free-tier user provides a `persona` in the request.
            Indicates that persona scoring requires a paid plan.
          example: "Persona scoring available on Solo plan ($29/mo)"
    ItemScore:
      type: object
      properties:
        title:
          type: string
          description: Title of the backlog item (matches RefinedItem.title)
        completeness_score:
          type: integer
          minimum: 0
          maximum: 100
          description: Completeness score from 0-100 indicating how ready the spec is for an AI coding agent
        agent_ready:
          type: boolean
          description: Whether the spec is ready for an AI coding agent to implement without clarification
        breakdown:
          type: object
          description: Boolean breakdown of which spec fields are present and complete
          properties:
            has_problem_statement:
              type: boolean
            has_acceptance_criteria:
              type: boolean
            has_estimate:
              type: boolean
            has_priority:
              type: boolean
            has_tags:
              type: boolean
    ScoreSummary:
      type: object
      description: Aggregate scoring summary across all items in the request
      properties:
        average_score:
          type: number
          description: Average completeness_score across all items
        agent_ready_count:
          type: integer
          description: Number of items with agent_ready = true
        total_count:
          type: integer
          description: Total number of items scored
    RefinedItem:
      type: object
      required:
        - title
        - problem
        - acceptanceCriteria
        - estimate
        - priority
        - tags
      properties:
        title:
          type: string
          description: Clean, actionable title for the backlog item
          example: "Fix authentication failure on login with special characters in password"
        problem:
          type: string
          description: Why this matters — concise problem statement (1-2 sentences)
          example: "Users with passwords containing special characters are unable to log in."
        acceptanceCriteria:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 4
          description: 2-4 testable acceptance criteria
          example:
            - "Users with passwords containing !@#$%^&*() can log in successfully"
            - "Error messages are clear and actionable when login fails"
        estimate:
          type: string
          enum: [XS, S, M, L, XL]
          description: |
            T-shirt size effort estimate:
            - XS: Less than 1 day
            - S: 1-2 days
            - M: 3-5 days
            - L: 1-2 weeks
            - XL: 2+ weeks
          example: "S"
        priority:
          type: string
          description: |
            Priority level with rationale. Format: "LEVEL — rationale"
            Levels: HIGH, MEDIUM, LOW
          example: "HIGH — causes account lockouts and direct user impact"
        tags:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 3
          description: 1-3 suggested labels/categories
          example: ["bug", "auth"]
        assumptions:
          type: array
          items:
            type: string
          description: |
            Optional. Open questions or assumptions that should be clarified
            before implementation. Only included when genuinely ambiguous.
          example:
            - "Assumes the issue is frontend input sanitization, not backend validation"
        persona_alignment:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          description: |
            Persona alignment score (0-100). Only present when a `persona` was provided
            in the request and the tier is Solo or higher. Null otherwise.
            Scoring: concern_coverage (50pts) + anti_concern_avoidance (25pts) + role_appropriate_language (25pts)
          example: 75
        persona_gaps:
          type: array
          items:
            type: string
          nullable: true
          description: |
            List of `cares_about` items from the persona that are NOT addressed in the spec.
            Only present when persona_alignment is a number. Null otherwise.
          example: ["CI integration"]
    ResponseMeta:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
          description: Unique identifier for this request
        model:
          type: string
          description: AI model used for refinement
          example: "claude-3-5-haiku-20241022"
        inputTokens:
          type: integer
          description: Number of input tokens consumed
        outputTokens:
          type: integer
          description: Number of output tokens generated
        costUsd:
          type: number
          format: float
          description: Estimated cost in USD for this request
        latencyMs:
          type: integer
          description: Request processing time in milliseconds
        promptVersion:
          type: string
          description: Version of the refinement prompt used
          example: "1.0"
        tier:
          type: string
          enum: [free, pro, team]
          description: Tier used for this request
...
    LintReceipt:
      type: object
      description: Stored lint receipt — verifiable proof that a spec was linted by Speclint
      properties:
        lint_id:
          type: string
          description: Unique receipt ID
          example: "spl_a1b2c3d4"
        score:
          type: integer
          description: Completeness score at time of linting
          example: 82
        breakdown:
          type: object
          description: Boolean breakdown of which spec fields were present
        title:
          type: string
          description: Title of the primary spec item that was linted
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the lint was performed
        tier:
          type: string
          description: License tier used for the lint request
          example: "free"
        agent_ready:
          type: boolean
          description: Whether the spec passed the agent-ready threshold at time of linting
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message